split project into modules core,client,server.
we produce 2 different artifacts and decouple things.
this way we no longer need isServer or isClient i.e. and we can separate server specific changes in to the server module.

configured Lwjgl to start without audio for the server module.
removed silentlogger, start with -Droot.log.level=ERROR -Dlog4j.configurationFile=none.xml
instead to silence even the startup noise from modthespire

based on previous patches
