+++ .gitignore
@@ -31,3 +31,4 @@ BattleAiMod.iml
 # maven
 target/
 dependency-reduced-pom.xml
+/_ModTheSpire/
