+++ .run/Client.run.xml
@@ -1,0 +1,13 @@
+<component name="ProjectRunConfigurationManager">
+  <configuration default="false" name="Client" type="JarApplication">
+    <option name="JAR_PATH" value="$PROJECT_DIR$/../_ModTheSpire/ModTheSpire.jar" />
+    <option name="VM_PARAMETERS" value="-Droot.log.level=ERROR -Dlog4j.configurationFile=none.xml" />
+    <option name="PROGRAM_PARAMETERS" value="--skip-intro --profile ai-client --mods basemod,stslib,CommunicationMod,LudicrousSpeed,SaveStateMod,client --skip-launcher" />
+    <option name="WORKING_DIRECTORY" value="$PROJECT_DIR$/../_ModTheSpire" />
+    <option name="ALTERNATIVE_JRE_PATH_ENABLED" value="true" />
+    <option name="ALTERNATIVE_JRE_PATH" value="1.8" />
+    <method v="2">
+      <option name="Maven.BeforeRunTask" enabled="true" file="$PROJECT_DIR$/pom.xml" goal="package --projects client,core" />
+    </method>
+  </configuration>
+</component>
