+++ client/src/main/resources/img/StartSteve.png
