+++ client/src/main/resources/img/fridge.png
