+++ core/src/main/java/autoplay/battleaimod/core/experimental/Images.java
@@ -1,0 +1,74 @@
+package autoplay.battleaimod.core.experimental;
+
+import com.badlogic.gdx.Gdx;
+import com.badlogic.gdx.files.FileHandle;
+import com.badlogic.gdx.graphics.Pixmap;
+import com.badlogic.gdx.graphics.PixmapIO;
+import com.badlogic.gdx.utils.ScreenUtils;
+
+import java.awt.*;
+import java.awt.image.BufferedImage;
+import java.io.IOException;
+import java.nio.ByteBuffer;
+
+public class Images {
+    static int screenshotCounter = 0;
+
+    private static BufferedImage resizeImage(BufferedImage originalImage, int targetWidth, int targetHeight) throws IOException {
+        int originaHeight = originalImage.getHeight();
+        int originalWidth = originalImage.getWidth();
+
+        double originalRatio = (double) originaHeight / (double) originalWidth;
+        double targetRatio = (double) targetHeight / (double) targetWidth;
+
+        int actualWidth;
+        int actualHeight;
+
+        if (originalRatio > targetRatio) {
+            // match height
+            actualHeight = Math.min(targetHeight, originaHeight);
+            actualWidth = (int) (actualHeight / originalRatio);
+        } else {
+            actualWidth = Math.min(targetWidth, originalWidth);
+            actualHeight = (int) (actualWidth * originalRatio);
+        }
+
+        BufferedImage resizedImage = new BufferedImage(actualWidth, actualHeight, BufferedImage.TYPE_INT_RGB);
+        Graphics2D graphics2D = resizedImage.createGraphics();
+        graphics2D.drawImage(originalImage, 0, 0, actualWidth, actualHeight, null);
+        graphics2D.dispose();
+        return resizedImage;
+    }
+
+    public static void takeScreenshot(String fileName) {
+        try {
+            FileHandle fh = new FileHandle(String
+                    .format("images\\%s_%s.png", fileName, screenshotCounter++));
+
+            Pixmap pixmap = getScreenshot(0, 0, Gdx.graphics.getWidth(), Gdx.graphics.getHeight(), true);
+            PixmapIO.writePNG(fh, pixmap);
+            pixmap.dispose();
+        } catch (Exception e) {
+        }
+    }
+
+    private static Pixmap getScreenshot(int x, int y, int w, int h, boolean yDown){
+        final Pixmap pixmap = ScreenUtils.getFrameBufferPixmap(x, y, w, h);
+
+        if (yDown) {
+            // Flip the pixmap upside down
+            ByteBuffer pixels = pixmap.getPixels();
+            int numBytes = w * h * 4;
+            byte[] lines = new byte[numBytes];
+            int numBytesPerLine = w * 4;
+            for (int i = 0; i < h; i++) {
+                pixels.position((h - i - 1) * numBytesPerLine);
+                pixels.get(lines, i * numBytesPerLine, numBytesPerLine);
+            }
+            pixels.clear();
+            pixels.put(lines);
+        }
+
+        return pixmap;
+    }
+}
