+++ core/src/main/java/autoplay/battleaimod/core/networking/Constants.java
@@ -1,0 +1,5 @@
+package autoplay.battleaimod.core.networking;
+
+public class Constants {
+    public static final int PORT_NUMBER = 5000;
+}
