+++ core/src/main/java/autoplay/battleaimod/core/networking/MessageType.java
@@ -1,0 +1,7 @@
+package autoplay.battleaimod.core.networking;
+
+public enum MessageType {
+    DONE,
+    STATUS_UPDATE,
+    COMMAND_LIST;
+}
