+++ core/src/main/resources/Icon.png
