--- src/main/java/battleaimod/SilentLogger.java
@@ -1,1868 +1,0 @@
-package battleaimod;
-
-import org.apache.logging.log4j.Level;
-import org.apache.logging.log4j.Logger;
-import org.apache.logging.log4j.Marker;
-import org.apache.logging.log4j.message.EntryMessage;
-import org.apache.logging.log4j.message.Message;
-import org.apache.logging.log4j.message.MessageFactory;
-import org.apache.logging.log4j.util.MessageSupplier;
-import org.apache.logging.log4j.util.Supplier;
-
-/**
- * A logger that doesn't do anything
- */
-public class SilentLogger implements Logger {
-    @Override
-    public void catching(Level level, Throwable throwable) {
-    }
-
-    @Override
-    public void catching(Throwable throwable) {
-    }
-
-    @Override
-    public void debug(Marker marker, Message message) {
-    }
-
-    @Override
-    public void debug(Marker marker, Message message, Throwable throwable) {
-    }
-
-    @Override
-    public void debug(Marker marker, MessageSupplier messageSupplier) {
-    }
-
-    @Override
-    public void debug(Marker marker, MessageSupplier messageSupplier, Throwable throwable) {
-    }
-
-    @Override
-    public void debug(Marker marker, CharSequence charSequence) {
-    }
-
-    @Override
-    public void debug(Marker marker, CharSequence charSequence, Throwable throwable) {
-    }
-
-    @Override
-    public void debug(Marker marker, Object o) {
-    }
-
-    @Override
-    public void debug(Marker marker, Object o, Throwable throwable) {
-    }
-
-    @Override
-    public void debug(Marker marker, String s) {
-    }
-
-    @Override
-    public void debug(Marker marker, String s, Object... objects) {
-    }
-
-    @Override
-    public void debug(Marker marker, String s, Supplier<?>... suppliers) {
-    }
-
-    @Override
-    public void debug(Marker marker, String s, Throwable throwable) {
-    }
-
-    @Override
-    public void debug(Marker marker, Supplier<?> supplier) {
-    }
-
-    @Override
-    public void debug(Marker marker, Supplier<?> supplier, Throwable throwable) {
-    }
-
-    @Override
-    public void debug(Message message) {
-    }
-
-    @Override
-    public void debug(Message message, Throwable throwable) {
-
-    }
-
-    @Override
-    public void debug(MessageSupplier messageSupplier) {
-
-    }
-
-    @Override
-    public void debug(MessageSupplier messageSupplier, Throwable throwable) {
-
-    }
-
-    @Override
-    public void debug(CharSequence charSequence) {
-
-    }
-
-    @Override
-    public void debug(CharSequence charSequence, Throwable throwable) {
-
-    }
-
-    @Override
-    public void debug(Object o) {
-
-    }
-
-    @Override
-    public void debug(Object o, Throwable throwable) {
-
-    }
-
-    @Override
-    public void debug(String s) {
-
-    }
-
-    @Override
-    public void debug(String s, Object... objects) {
-
-    }
-
-    @Override
-    public void debug(String s, Supplier<?>... suppliers) {
-
-    }
-
-    @Override
-    public void debug(String s, Throwable throwable) {
-
-    }
-
-    @Override
-    public void debug(Supplier<?> supplier) {
-
-    }
-
-    @Override
-    public void debug(Supplier<?> supplier, Throwable throwable) {
-
-    }
-
-    @Override
-    public void debug(Marker marker, String s, Object o) {
-
-    }
-
-    @Override
-    public void debug(Marker marker, String s, Object o, Object o1) {
-
-    }
-
-    @Override
-    public void debug(Marker marker, String s, Object o, Object o1, Object o2) {
-
-    }
-
-    @Override
-    public void debug(Marker marker, String s, Object o, Object o1, Object o2, Object o3) {
-
-    }
-
-    @Override
-    public void debug(Marker marker, String s, Object o, Object o1, Object o2, Object o3, Object o4) {
-
-    }
-
-    @Override
-    public void debug(Marker marker, String s, Object o, Object o1, Object o2, Object o3, Object o4, Object o5) {
-
-    }
-
-    @Override
-    public void debug(Marker marker, String s, Object o, Object o1, Object o2, Object o3, Object o4, Object o5, Object o6) {
-
-    }
-
-    @Override
-    public void debug(Marker marker, String s, Object o, Object o1, Object o2, Object o3, Object o4, Object o5, Object o6, Object o7) {
-
-    }
-
-    @Override
-    public void debug(Marker marker, String s, Object o, Object o1, Object o2, Object o3, Object o4, Object o5, Object o6, Object o7, Object o8) {
-
-    }
-
-    @Override
-    public void debug(Marker marker, String s, Object o, Object o1, Object o2, Object o3, Object o4, Object o5, Object o6, Object o7, Object o8, Object o9) {
-
-    }
-
-    @Override
-    public void debug(String s, Object o) {
-
-    }
-
-    @Override
-    public void debug(String s, Object o, Object o1) {
-
-    }
-
-    @Override
-    public void debug(String s, Object o, Object o1, Object o2) {
-
-    }
-
-    @Override
-    public void debug(String s, Object o, Object o1, Object o2, Object o3) {
-
-    }
-
-    @Override
-    public void debug(String s, Object o, Object o1, Object o2, Object o3, Object o4) {
-
-    }
-
-    @Override
-    public void debug(String s, Object o, Object o1, Object o2, Object o3, Object o4, Object o5) {
-
-    }
-
-    @Override
-    public void debug(String s, Object o, Object o1, Object o2, Object o3, Object o4, Object o5, Object o6) {
-
-    }
-
-    @Override
-    public void debug(String s, Object o, Object o1, Object o2, Object o3, Object o4, Object o5, Object o6, Object o7) {
-
-    }
-
-    @Override
-    public void debug(String s, Object o, Object o1, Object o2, Object o3, Object o4, Object o5, Object o6, Object o7, Object o8) {
-
-    }
-
-    @Override
-    public void debug(String s, Object o, Object o1, Object o2, Object o3, Object o4, Object o5, Object o6, Object o7, Object o8, Object o9) {
-
-    }
-
-    @Override
-    public void entry() {
-
-    }
-
-    @Override
-    public void entry(Object... objects) {
-
-    }
-
-    @Override
-    public void error(Marker marker, Message message) {
-
-    }
-
-    @Override
-    public void error(Marker marker, Message message, Throwable throwable) {
-
-    }
-
-    @Override
-    public void error(Marker marker, MessageSupplier messageSupplier) {
-
-    }
-
-    @Override
-    public void error(Marker marker, MessageSupplier messageSupplier, Throwable throwable) {
-
-    }
-
-    @Override
-    public void error(Marker marker, CharSequence charSequence) {
-
-    }
-
-    @Override
-    public void error(Marker marker, CharSequence charSequence, Throwable throwable) {
-
-    }
-
-    @Override
-    public void error(Marker marker, Object o) {
-
-    }
-
-    @Override
-    public void error(Marker marker, Object o, Throwable throwable) {
-
-    }
-
-    @Override
-    public void error(Marker marker, String s) {
-
-    }
-
-    @Override
-    public void error(Marker marker, String s, Object... objects) {
-
-    }
-
-    @Override
-    public void error(Marker marker, String s, Supplier<?>... suppliers) {
-
-    }
-
-    @Override
-    public void error(Marker marker, String s, Throwable throwable) {
-
-    }
-
-    @Override
-    public void error(Marker marker, Supplier<?> supplier) {
-
-    }
-
-    @Override
-    public void error(Marker marker, Supplier<?> supplier, Throwable throwable) {
-
-    }
-
-    @Override
-    public void error(Message message) {
-
-    }
-
-    @Override
-    public void error(Message message, Throwable throwable) {
-
-    }
-
-    @Override
-    public void error(MessageSupplier messageSupplier) {
-
-    }
-
-    @Override
-    public void error(MessageSupplier messageSupplier, Throwable throwable) {
-
-    }
-
-    @Override
-    public void error(CharSequence charSequence) {
-
-    }
-
-    @Override
-    public void error(CharSequence charSequence, Throwable throwable) {
-
-    }
-
-    @Override
-    public void error(Object o) {
-
-    }
-
-    @Override
-    public void error(Object o, Throwable throwable) {
-
-    }
-
-    @Override
-    public void error(String s) {
-
-    }
-
-    @Override
-    public void error(String s, Object... objects) {
-
-    }
-
-    @Override
-    public void error(String s, Supplier<?>... suppliers) {
-
-    }
-
-    @Override
-    public void error(String s, Throwable throwable) {
-
-    }
-
-    @Override
-    public void error(Supplier<?> supplier) {
-
-    }
-
-    @Override
-    public void error(Supplier<?> supplier, Throwable throwable) {
-
-    }
-
-    @Override
-    public void error(Marker marker, String s, Object o) {
-
-    }
-
-    @Override
-    public void error(Marker marker, String s, Object o, Object o1) {
-
-    }
-
-    @Override
-    public void error(Marker marker, String s, Object o, Object o1, Object o2) {
-
-    }
-
-    @Override
-    public void error(Marker marker, String s, Object o, Object o1, Object o2, Object o3) {
-
-    }
-
-    @Override
-    public void error(Marker marker, String s, Object o, Object o1, Object o2, Object o3, Object o4) {
-
-    }
-
-    @Override
-    public void error(Marker marker, String s, Object o, Object o1, Object o2, Object o3, Object o4, Object o5) {
-
-    }
-
-    @Override
-    public void error(Marker marker, String s, Object o, Object o1, Object o2, Object o3, Object o4, Object o5, Object o6) {
-
-    }
-
-    @Override
-    public void error(Marker marker, String s, Object o, Object o1, Object o2, Object o3, Object o4, Object o5, Object o6, Object o7) {
-
-    }
-
-    @Override
-    public void error(Marker marker, String s, Object o, Object o1, Object o2, Object o3, Object o4, Object o5, Object o6, Object o7, Object o8) {
-
-    }
-
-    @Override
-    public void error(Marker marker, String s, Object o, Object o1, Object o2, Object o3, Object o4, Object o5, Object o6, Object o7, Object o8, Object o9) {
-
-    }
-
-    @Override
-    public void error(String s, Object o) {
-
-    }
-
-    @Override
-    public void error(String s, Object o, Object o1) {
-
-    }
-
-    @Override
-    public void error(String s, Object o, Object o1, Object o2) {
-
-    }
-
-    @Override
-    public void error(String s, Object o, Object o1, Object o2, Object o3) {
-
-    }
-
-    @Override
-    public void error(String s, Object o, Object o1, Object o2, Object o3, Object o4) {
-
-    }
-
-    @Override
-    public void error(String s, Object o, Object o1, Object o2, Object o3, Object o4, Object o5) {
-
-    }
-
-    @Override
-    public void error(String s, Object o, Object o1, Object o2, Object o3, Object o4, Object o5, Object o6) {
-
-    }
-
-    @Override
-    public void error(String s, Object o, Object o1, Object o2, Object o3, Object o4, Object o5, Object o6, Object o7) {
-
-    }
-
-    @Override
-    public void error(String s, Object o, Object o1, Object o2, Object o3, Object o4, Object o5, Object o6, Object o7, Object o8) {
-
-    }
-
-    @Override
-    public void error(String s, Object o, Object o1, Object o2, Object o3, Object o4, Object o5, Object o6, Object o7, Object o8, Object o9) {
-
-    }
-
-    @Override
-    public void exit() {
-
-    }
-
-    @Override
-    public <R> R exit(R r) {
-        return null;
-    }
-
-    @Override
-    public void fatal(Marker marker, Message message) {
-
-    }
-
-    @Override
-    public void fatal(Marker marker, Message message, Throwable throwable) {
-
-    }
-
-    @Override
-    public void fatal(Marker marker, MessageSupplier messageSupplier) {
-
-    }
-
-    @Override
-    public void fatal(Marker marker, MessageSupplier messageSupplier, Throwable throwable) {
-
-    }
-
-    @Override
-    public void fatal(Marker marker, CharSequence charSequence) {
-
-    }
-
-    @Override
-    public void fatal(Marker marker, CharSequence charSequence, Throwable throwable) {
-
-    }
-
-    @Override
-    public void fatal(Marker marker, Object o) {
-
-    }
-
-    @Override
-    public void fatal(Marker marker, Object o, Throwable throwable) {
-
-    }
-
-    @Override
-    public void fatal(Marker marker, String s) {
-
-    }
-
-    @Override
-    public void fatal(Marker marker, String s, Object... objects) {
-
-    }
-
-    @Override
-    public void fatal(Marker marker, String s, Supplier<?>... suppliers) {
-
-    }
-
-    @Override
-    public void fatal(Marker marker, String s, Throwable throwable) {
-
-    }
-
-    @Override
-    public void fatal(Marker marker, Supplier<?> supplier) {
-
-    }
-
-    @Override
-    public void fatal(Marker marker, Supplier<?> supplier, Throwable throwable) {
-
-    }
-
-    @Override
-    public void fatal(Message message) {
-
-    }
-
-    @Override
-    public void fatal(Message message, Throwable throwable) {
-
-    }
-
-    @Override
-    public void fatal(MessageSupplier messageSupplier) {
-
-    }
-
-    @Override
-    public void fatal(MessageSupplier messageSupplier, Throwable throwable) {
-
-    }
-
-    @Override
-    public void fatal(CharSequence charSequence) {
-
-    }
-
-    @Override
-    public void fatal(CharSequence charSequence, Throwable throwable) {
-
-    }
-
-    @Override
-    public void fatal(Object o) {
-
-    }
-
-    @Override
-    public void fatal(Object o, Throwable throwable) {
-
-    }
-
-    @Override
-    public void fatal(String s) {
-
-    }
-
-    @Override
-    public void fatal(String s, Object... objects) {
-
-    }
-
-    @Override
-    public void fatal(String s, Supplier<?>... suppliers) {
-
-    }
-
-    @Override
-    public void fatal(String s, Throwable throwable) {
-
-    }
-
-    @Override
-    public void fatal(Supplier<?> supplier) {
-
-    }
-
-    @Override
-    public void fatal(Supplier<?> supplier, Throwable throwable) {
-
-    }
-
-    @Override
-    public void fatal(Marker marker, String s, Object o) {
-
-    }
-
-    @Override
-    public void fatal(Marker marker, String s, Object o, Object o1) {
-
-    }
-
-    @Override
-    public void fatal(Marker marker, String s, Object o, Object o1, Object o2) {
-
-    }
-
-    @Override
-    public void fatal(Marker marker, String s, Object o, Object o1, Object o2, Object o3) {
-
-    }
-
-    @Override
-    public void fatal(Marker marker, String s, Object o, Object o1, Object o2, Object o3, Object o4) {
-
-    }
-
-    @Override
-    public void fatal(Marker marker, String s, Object o, Object o1, Object o2, Object o3, Object o4, Object o5) {
-
-    }
-
-    @Override
-    public void fatal(Marker marker, String s, Object o, Object o1, Object o2, Object o3, Object o4, Object o5, Object o6) {
-
-    }
-
-    @Override
-    public void fatal(Marker marker, String s, Object o, Object o1, Object o2, Object o3, Object o4, Object o5, Object o6, Object o7) {
-
-    }
-
-    @Override
-    public void fatal(Marker marker, String s, Object o, Object o1, Object o2, Object o3, Object o4, Object o5, Object o6, Object o7, Object o8) {
-
-    }
-
-    @Override
-    public void fatal(Marker marker, String s, Object o, Object o1, Object o2, Object o3, Object o4, Object o5, Object o6, Object o7, Object o8, Object o9) {
-
-    }
-
-    @Override
-    public void fatal(String s, Object o) {
-
-    }
-
-    @Override
-    public void fatal(String s, Object o, Object o1) {
-
-    }
-
-    @Override
-    public void fatal(String s, Object o, Object o1, Object o2) {
-
-    }
-
-    @Override
-    public void fatal(String s, Object o, Object o1, Object o2, Object o3) {
-
-    }
-
-    @Override
-    public void fatal(String s, Object o, Object o1, Object o2, Object o3, Object o4) {
-
-    }
-
-    @Override
-    public void fatal(String s, Object o, Object o1, Object o2, Object o3, Object o4, Object o5) {
-
-    }
-
-    @Override
-    public void fatal(String s, Object o, Object o1, Object o2, Object o3, Object o4, Object o5, Object o6) {
-
-    }
-
-    @Override
-    public void fatal(String s, Object o, Object o1, Object o2, Object o3, Object o4, Object o5, Object o6, Object o7) {
-
-    }
-
-    @Override
-    public void fatal(String s, Object o, Object o1, Object o2, Object o3, Object o4, Object o5, Object o6, Object o7, Object o8) {
-
-    }
-
-    @Override
-    public void fatal(String s, Object o, Object o1, Object o2, Object o3, Object o4, Object o5, Object o6, Object o7, Object o8, Object o9) {
-
-    }
-
-    @Override
-    public Level getLevel() {
-        return null;
-    }
-
-    @Override
-    public <MF extends MessageFactory> MF getMessageFactory() {
-        return null;
-    }
-
-    @Override
-    public String getName() {
-        return null;
-    }
-
-    @Override
-    public void info(Marker marker, Message message) {
-
-    }
-
-    @Override
-    public void info(Marker marker, Message message, Throwable throwable) {
-
-    }
-
-    @Override
-    public void info(Marker marker, MessageSupplier messageSupplier) {
-
-    }
-
-    @Override
-    public void info(Marker marker, MessageSupplier messageSupplier, Throwable throwable) {
-
-    }
-
-    @Override
-    public void info(Marker marker, CharSequence charSequence) {
-
-    }
-
-    @Override
-    public void info(Marker marker, CharSequence charSequence, Throwable throwable) {
-
-    }
-
-    @Override
-    public void info(Marker marker, Object o) {
-
-    }
-
-    @Override
-    public void info(Marker marker, Object o, Throwable throwable) {
-
-    }
-
-    @Override
-    public void info(Marker marker, String s) {
-
-    }
-
-    @Override
-    public void info(Marker marker, String s, Object... objects) {
-
-    }
-
-    @Override
-    public void info(Marker marker, String s, Supplier<?>... suppliers) {
-
-    }
-
-    @Override
-    public void info(Marker marker, String s, Throwable throwable) {
-
-    }
-
-    @Override
-    public void info(Marker marker, Supplier<?> supplier) {
-
-    }
-
-    @Override
-    public void info(Marker marker, Supplier<?> supplier, Throwable throwable) {
-
-    }
-
-    @Override
-    public void info(Message message) {
-
-    }
-
-    @Override
-    public void info(Message message, Throwable throwable) {
-
-    }
-
-    @Override
-    public void info(MessageSupplier messageSupplier) {
-
-    }
-
-    @Override
-    public void info(MessageSupplier messageSupplier, Throwable throwable) {
-
-    }
-
-    @Override
-    public void info(CharSequence charSequence) {
-
-    }
-
-    @Override
-    public void info(CharSequence charSequence, Throwable throwable) {
-
-    }
-
-    @Override
-    public void info(Object o) {
-
-    }
-
-    @Override
-    public void info(Object o, Throwable throwable) {
-
-    }
-
-    @Override
-    public void info(String s) {
-
-    }
-
-    @Override
-    public void info(String s, Object... objects) {
-
-    }
-
-    @Override
-    public void info(String s, Supplier<?>... suppliers) {
-
-    }
-
-    @Override
-    public void info(String s, Throwable throwable) {
-
-    }
-
-    @Override
-    public void info(Supplier<?> supplier) {
-
-    }
-
-    @Override
-    public void info(Supplier<?> supplier, Throwable throwable) {
-
-    }
-
-    @Override
-    public void info(Marker marker, String s, Object o) {
-
-    }
-
-    @Override
-    public void info(Marker marker, String s, Object o, Object o1) {
-
-    }
-
-    @Override
-    public void info(Marker marker, String s, Object o, Object o1, Object o2) {
-
-    }
-
-    @Override
-    public void info(Marker marker, String s, Object o, Object o1, Object o2, Object o3) {
-
-    }
-
-    @Override
-    public void info(Marker marker, String s, Object o, Object o1, Object o2, Object o3, Object o4) {
-
-    }
-
-    @Override
-    public void info(Marker marker, String s, Object o, Object o1, Object o2, Object o3, Object o4, Object o5) {
-
-    }
-
-    @Override
-    public void info(Marker marker, String s, Object o, Object o1, Object o2, Object o3, Object o4, Object o5, Object o6) {
-
-    }
-
-    @Override
-    public void info(Marker marker, String s, Object o, Object o1, Object o2, Object o3, Object o4, Object o5, Object o6, Object o7) {
-
-    }
-
-    @Override
-    public void info(Marker marker, String s, Object o, Object o1, Object o2, Object o3, Object o4, Object o5, Object o6, Object o7, Object o8) {
-
-    }
-
-    @Override
-    public void info(Marker marker, String s, Object o, Object o1, Object o2, Object o3, Object o4, Object o5, Object o6, Object o7, Object o8, Object o9) {
-
-    }
-
-    @Override
-    public void info(String s, Object o) {
-
-    }
-
-    @Override
-    public void info(String s, Object o, Object o1) {
-
-    }
-
-    @Override
-    public void info(String s, Object o, Object o1, Object o2) {
-
-    }
-
-    @Override
-    public void info(String s, Object o, Object o1, Object o2, Object o3) {
-
-    }
-
-    @Override
-    public void info(String s, Object o, Object o1, Object o2, Object o3, Object o4) {
-
-    }
-
-    @Override
-    public void info(String s, Object o, Object o1, Object o2, Object o3, Object o4, Object o5) {
-
-    }
-
-    @Override
-    public void info(String s, Object o, Object o1, Object o2, Object o3, Object o4, Object o5, Object o6) {
-
-    }
-
-    @Override
-    public void info(String s, Object o, Object o1, Object o2, Object o3, Object o4, Object o5, Object o6, Object o7) {
-
-    }
-
-    @Override
-    public void info(String s, Object o, Object o1, Object o2, Object o3, Object o4, Object o5, Object o6, Object o7, Object o8) {
-
-    }
-
-    @Override
-    public void info(String s, Object o, Object o1, Object o2, Object o3, Object o4, Object o5, Object o6, Object o7, Object o8, Object o9) {
-
-    }
-
-    @Override
-    public boolean isDebugEnabled() {
-        return false;
-    }
-
-    @Override
-    public boolean isDebugEnabled(Marker marker) {
-        return false;
-    }
-
-    @Override
-    public boolean isEnabled(Level level) {
-        return false;
-    }
-
-    @Override
-    public boolean isEnabled(Level level, Marker marker) {
-        return false;
-    }
-
-    @Override
-    public boolean isErrorEnabled() {
-        return false;
-    }
-
-    @Override
-    public boolean isErrorEnabled(Marker marker) {
-        return false;
-    }
-
-    @Override
-    public boolean isFatalEnabled() {
-        return false;
-    }
-
-    @Override
-    public boolean isFatalEnabled(Marker marker) {
-        return false;
-    }
-
-    @Override
-    public boolean isInfoEnabled() {
-        return false;
-    }
-
-    @Override
-    public boolean isInfoEnabled(Marker marker) {
-        return false;
-    }
-
-    @Override
-    public boolean isTraceEnabled() {
-        return false;
-    }
-
-    @Override
-    public boolean isTraceEnabled(Marker marker) {
-        return false;
-    }
-
-    @Override
-    public boolean isWarnEnabled() {
-        return false;
-    }
-
-    @Override
-    public boolean isWarnEnabled(Marker marker) {
-        return false;
-    }
-
-    @Override
-    public void log(Level level, Marker marker, Message message) {
-
-    }
-
-    @Override
-    public void log(Level level, Marker marker, Message message, Throwable throwable) {
-
-    }
-
-    @Override
-    public void log(Level level, Marker marker, MessageSupplier messageSupplier) {
-
-    }
-
-    @Override
-    public void log(Level level, Marker marker, MessageSupplier messageSupplier, Throwable throwable) {
-
-    }
-
-    @Override
-    public void log(Level level, Marker marker, CharSequence charSequence) {
-
-    }
-
-    @Override
-    public void log(Level level, Marker marker, CharSequence charSequence, Throwable throwable) {
-
-    }
-
-    @Override
-    public void log(Level level, Marker marker, Object o) {
-
-    }
-
-    @Override
-    public void log(Level level, Marker marker, Object o, Throwable throwable) {
-
-    }
-
-    @Override
-    public void log(Level level, Marker marker, String s) {
-
-    }
-
-    @Override
-    public void log(Level level, Marker marker, String s, Object... objects) {
-
-    }
-
-    @Override
-    public void log(Level level, Marker marker, String s, Supplier<?>... suppliers) {
-
-    }
-
-    @Override
-    public void log(Level level, Marker marker, String s, Throwable throwable) {
-
-    }
-
-    @Override
-    public void log(Level level, Marker marker, Supplier<?> supplier) {
-
-    }
-
-    @Override
-    public void log(Level level, Marker marker, Supplier<?> supplier, Throwable throwable) {
-
-    }
-
-    @Override
-    public void log(Level level, Message message) {
-
-    }
-
-    @Override
-    public void log(Level level, Message message, Throwable throwable) {
-
-    }
-
-    @Override
-    public void log(Level level, MessageSupplier messageSupplier) {
-
-    }
-
-    @Override
-    public void log(Level level, MessageSupplier messageSupplier, Throwable throwable) {
-
-    }
-
-    @Override
-    public void log(Level level, CharSequence charSequence) {
-
-    }
-
-    @Override
-    public void log(Level level, CharSequence charSequence, Throwable throwable) {
-
-    }
-
-    @Override
-    public void log(Level level, Object o) {
-
-    }
-
-    @Override
-    public void log(Level level, Object o, Throwable throwable) {
-
-    }
-
-    @Override
-    public void log(Level level, String s) {
-
-    }
-
-    @Override
-    public void log(Level level, String s, Object... objects) {
-
-    }
-
-    @Override
-    public void log(Level level, String s, Supplier<?>... suppliers) {
-
-    }
-
-    @Override
-    public void log(Level level, String s, Throwable throwable) {
-
-    }
-
-    @Override
-    public void log(Level level, Supplier<?> supplier) {
-
-    }
-
-    @Override
-    public void log(Level level, Supplier<?> supplier, Throwable throwable) {
-
-    }
-
-    @Override
-    public void log(Level level, Marker marker, String s, Object o) {
-
-    }
-
-    @Override
-    public void log(Level level, Marker marker, String s, Object o, Object o1) {
-
-    }
-
-    @Override
-    public void log(Level level, Marker marker, String s, Object o, Object o1, Object o2) {
-
-    }
-
-    @Override
-    public void log(Level level, Marker marker, String s, Object o, Object o1, Object o2, Object o3) {
-
-    }
-
-    @Override
-    public void log(Level level, Marker marker, String s, Object o, Object o1, Object o2, Object o3, Object o4) {
-
-    }
-
-    @Override
-    public void log(Level level, Marker marker, String s, Object o, Object o1, Object o2, Object o3, Object o4, Object o5) {
-
-    }
-
-    @Override
-    public void log(Level level, Marker marker, String s, Object o, Object o1, Object o2, Object o3, Object o4, Object o5, Object o6) {
-
-    }
-
-    @Override
-    public void log(Level level, Marker marker, String s, Object o, Object o1, Object o2, Object o3, Object o4, Object o5, Object o6, Object o7) {
-
-    }
-
-    @Override
-    public void log(Level level, Marker marker, String s, Object o, Object o1, Object o2, Object o3, Object o4, Object o5, Object o6, Object o7, Object o8) {
-
-    }
-
-    @Override
-    public void log(Level level, Marker marker, String s, Object o, Object o1, Object o2, Object o3, Object o4, Object o5, Object o6, Object o7, Object o8, Object o9) {
-
-    }
-
-    @Override
-    public void log(Level level, String s, Object o) {
-
-    }
-
-    @Override
-    public void log(Level level, String s, Object o, Object o1) {
-
-    }
-
-    @Override
-    public void log(Level level, String s, Object o, Object o1, Object o2) {
-
-    }
-
-    @Override
-    public void log(Level level, String s, Object o, Object o1, Object o2, Object o3) {
-
-    }
-
-    @Override
-    public void log(Level level, String s, Object o, Object o1, Object o2, Object o3, Object o4) {
-
-    }
-
-    @Override
-    public void log(Level level, String s, Object o, Object o1, Object o2, Object o3, Object o4, Object o5) {
-
-    }
-
-    @Override
-    public void log(Level level, String s, Object o, Object o1, Object o2, Object o3, Object o4, Object o5, Object o6) {
-
-    }
-
-    @Override
-    public void log(Level level, String s, Object o, Object o1, Object o2, Object o3, Object o4, Object o5, Object o6, Object o7) {
-
-    }
-
-    @Override
-    public void log(Level level, String s, Object o, Object o1, Object o2, Object o3, Object o4, Object o5, Object o6, Object o7, Object o8) {
-
-    }
-
-    @Override
-    public void log(Level level, String s, Object o, Object o1, Object o2, Object o3, Object o4, Object o5, Object o6, Object o7, Object o8, Object o9) {
-
-    }
-
-    @Override
-    public void printf(Level level, Marker marker, String s, Object... objects) {
-
-    }
-
-    @Override
-    public void printf(Level level, String s, Object... objects) {
-
-    }
-
-    @Override
-    public <T extends Throwable> T throwing(Level level, T t) {
-        return null;
-    }
-
-    @Override
-    public <T extends Throwable> T throwing(T t) {
-        return null;
-    }
-
-    @Override
-    public void trace(Marker marker, Message message) {
-
-    }
-
-    @Override
-    public void trace(Marker marker, Message message, Throwable throwable) {
-
-    }
-
-    @Override
-    public void trace(Marker marker, MessageSupplier messageSupplier) {
-
-    }
-
-    @Override
-    public void trace(Marker marker, MessageSupplier messageSupplier, Throwable throwable) {
-
-    }
-
-    @Override
-    public void trace(Marker marker, CharSequence charSequence) {
-
-    }
-
-    @Override
-    public void trace(Marker marker, CharSequence charSequence, Throwable throwable) {
-
-    }
-
-    @Override
-    public void trace(Marker marker, Object o) {
-
-    }
-
-    @Override
-    public void trace(Marker marker, Object o, Throwable throwable) {
-
-    }
-
-    @Override
-    public void trace(Marker marker, String s) {
-
-    }
-
-    @Override
-    public void trace(Marker marker, String s, Object... objects) {
-
-    }
-
-    @Override
-    public void trace(Marker marker, String s, Supplier<?>... suppliers) {
-
-    }
-
-    @Override
-    public void trace(Marker marker, String s, Throwable throwable) {
-
-    }
-
-    @Override
-    public void trace(Marker marker, Supplier<?> supplier) {
-
-    }
-
-    @Override
-    public void trace(Marker marker, Supplier<?> supplier, Throwable throwable) {
-
-    }
-
-    @Override
-    public void trace(Message message) {
-
-    }
-
-    @Override
-    public void trace(Message message, Throwable throwable) {
-
-    }
-
-    @Override
-    public void trace(MessageSupplier messageSupplier) {
-
-    }
-
-    @Override
-    public void trace(MessageSupplier messageSupplier, Throwable throwable) {
-
-    }
-
-    @Override
-    public void trace(CharSequence charSequence) {
-
-    }
-
-    @Override
-    public void trace(CharSequence charSequence, Throwable throwable) {
-
-    }
-
-    @Override
-    public void trace(Object o) {
-
-    }
-
-    @Override
-    public void trace(Object o, Throwable throwable) {
-
-    }
-
-    @Override
-    public void trace(String s) {
-
-    }
-
-    @Override
-    public void trace(String s, Object... objects) {
-
-    }
-
-    @Override
-    public void trace(String s, Supplier<?>... suppliers) {
-
-    }
-
-    @Override
-    public void trace(String s, Throwable throwable) {
-
-    }
-
-    @Override
-    public void trace(Supplier<?> supplier) {
-
-    }
-
-    @Override
-    public void trace(Supplier<?> supplier, Throwable throwable) {
-
-    }
-
-    @Override
-    public void trace(Marker marker, String s, Object o) {
-
-    }
-
-    @Override
-    public void trace(Marker marker, String s, Object o, Object o1) {
-
-    }
-
-    @Override
-    public void trace(Marker marker, String s, Object o, Object o1, Object o2) {
-
-    }
-
-    @Override
-    public void trace(Marker marker, String s, Object o, Object o1, Object o2, Object o3) {
-
-    }
-
-    @Override
-    public void trace(Marker marker, String s, Object o, Object o1, Object o2, Object o3, Object o4) {
-
-    }
-
-    @Override
-    public void trace(Marker marker, String s, Object o, Object o1, Object o2, Object o3, Object o4, Object o5) {
-
-    }
-
-    @Override
-    public void trace(Marker marker, String s, Object o, Object o1, Object o2, Object o3, Object o4, Object o5, Object o6) {
-
-    }
-
-    @Override
-    public void trace(Marker marker, String s, Object o, Object o1, Object o2, Object o3, Object o4, Object o5, Object o6, Object o7) {
-
-    }
-
-    @Override
-    public void trace(Marker marker, String s, Object o, Object o1, Object o2, Object o3, Object o4, Object o5, Object o6, Object o7, Object o8) {
-
-    }
-
-    @Override
-    public void trace(Marker marker, String s, Object o, Object o1, Object o2, Object o3, Object o4, Object o5, Object o6, Object o7, Object o8, Object o9) {
-
-    }
-
-    @Override
-    public void trace(String s, Object o) {
-
-    }
-
-    @Override
-    public void trace(String s, Object o, Object o1) {
-
-    }
-
-    @Override
-    public void trace(String s, Object o, Object o1, Object o2) {
-
-    }
-
-    @Override
-    public void trace(String s, Object o, Object o1, Object o2, Object o3) {
-
-    }
-
-    @Override
-    public void trace(String s, Object o, Object o1, Object o2, Object o3, Object o4) {
-
-    }
-
-    @Override
-    public void trace(String s, Object o, Object o1, Object o2, Object o3, Object o4, Object o5) {
-
-    }
-
-    @Override
-    public void trace(String s, Object o, Object o1, Object o2, Object o3, Object o4, Object o5, Object o6) {
-
-    }
-
-    @Override
-    public void trace(String s, Object o, Object o1, Object o2, Object o3, Object o4, Object o5, Object o6, Object o7) {
-
-    }
-
-    @Override
-    public void trace(String s, Object o, Object o1, Object o2, Object o3, Object o4, Object o5, Object o6, Object o7, Object o8) {
-
-    }
-
-    @Override
-    public void trace(String s, Object o, Object o1, Object o2, Object o3, Object o4, Object o5, Object o6, Object o7, Object o8, Object o9) {
-
-    }
-
-    @Override
-    public EntryMessage traceEntry() {
-        return null;
-    }
-
-    @Override
-    public EntryMessage traceEntry(String s, Object... objects) {
-        return null;
-    }
-
-    @Override
-    public EntryMessage traceEntry(Supplier<?>... suppliers) {
-        return null;
-    }
-
-    @Override
-    public EntryMessage traceEntry(String s, Supplier<?>... suppliers) {
-        return null;
-    }
-
-    @Override
-    public EntryMessage traceEntry(Message message) {
-        return null;
-    }
-
-    @Override
-    public void traceExit() {
-
-    }
-
-    @Override
-    public <R> R traceExit(R r) {
-        return null;
-    }
-
-    @Override
-    public <R> R traceExit(String s, R r) {
-        return null;
-    }
-
-    @Override
-    public void traceExit(EntryMessage entryMessage) {
-
-    }
-
-    @Override
-    public <R> R traceExit(EntryMessage entryMessage, R r) {
-        return null;
-    }
-
-    @Override
-    public <R> R traceExit(Message message, R r) {
-        return null;
-    }
-
-    @Override
-    public void warn(Marker marker, Message message) {
-
-    }
-
-    @Override
-    public void warn(Marker marker, Message message, Throwable throwable) {
-
-    }
-
-    @Override
-    public void warn(Marker marker, MessageSupplier messageSupplier) {
-
-    }
-
-    @Override
-    public void warn(Marker marker, MessageSupplier messageSupplier, Throwable throwable) {
-
-    }
-
-    @Override
-    public void warn(Marker marker, CharSequence charSequence) {
-
-    }
-
-    @Override
-    public void warn(Marker marker, CharSequence charSequence, Throwable throwable) {
-
-    }
-
-    @Override
-    public void warn(Marker marker, Object o) {
-
-    }
-
-    @Override
-    public void warn(Marker marker, Object o, Throwable throwable) {
-
-    }
-
-    @Override
-    public void warn(Marker marker, String s) {
-
-    }
-
-    @Override
-    public void warn(Marker marker, String s, Object... objects) {
-
-    }
-
-    @Override
-    public void warn(Marker marker, String s, Supplier<?>... suppliers) {
-
-    }
-
-    @Override
-    public void warn(Marker marker, String s, Throwable throwable) {
-
-    }
-
-    @Override
-    public void warn(Marker marker, Supplier<?> supplier) {
-
-    }
-
-    @Override
-    public void warn(Marker marker, Supplier<?> supplier, Throwable throwable) {
-
-    }
-
-    @Override
-    public void warn(Message message) {
-
-    }
-
-    @Override
-    public void warn(Message message, Throwable throwable) {
-
-    }
-
-    @Override
-    public void warn(MessageSupplier messageSupplier) {
-
-    }
-
-    @Override
-    public void warn(MessageSupplier messageSupplier, Throwable throwable) {
-
-    }
-
-    @Override
-    public void warn(CharSequence charSequence) {
-
-    }
-
-    @Override
-    public void warn(CharSequence charSequence, Throwable throwable) {
-
-    }
-
-    @Override
-    public void warn(Object o) {
-
-    }
-
-    @Override
-    public void warn(Object o, Throwable throwable) {
-
-    }
-
-    @Override
-    public void warn(String s) {
-
-    }
-
-    @Override
-    public void warn(String s, Object... objects) {
-
-    }
-
-    @Override
-    public void warn(String s, Supplier<?>... suppliers) {
-
-    }
-
-    @Override
-    public void warn(String s, Throwable throwable) {
-
-    }
-
-    @Override
-    public void warn(Supplier<?> supplier) {
-
-    }
-
-    @Override
-    public void warn(Supplier<?> supplier, Throwable throwable) {
-
-    }
-
-    @Override
-    public void warn(Marker marker, String s, Object o) {
-
-    }
-
-    @Override
-    public void warn(Marker marker, String s, Object o, Object o1) {
-
-    }
-
-    @Override
-    public void warn(Marker marker, String s, Object o, Object o1, Object o2) {
-
-    }
-
-    @Override
-    public void warn(Marker marker, String s, Object o, Object o1, Object o2, Object o3) {
-
-    }
-
-    @Override
-    public void warn(Marker marker, String s, Object o, Object o1, Object o2, Object o3, Object o4) {
-
-    }
-
-    @Override
-    public void warn(Marker marker, String s, Object o, Object o1, Object o2, Object o3, Object o4, Object o5) {
-
-    }
-
-    @Override
-    public void warn(Marker marker, String s, Object o, Object o1, Object o2, Object o3, Object o4, Object o5, Object o6) {
-
-    }
-
-    @Override
-    public void warn(Marker marker, String s, Object o, Object o1, Object o2, Object o3, Object o4, Object o5, Object o6, Object o7) {
-
-    }
-
-    @Override
-    public void warn(Marker marker, String s, Object o, Object o1, Object o2, Object o3, Object o4, Object o5, Object o6, Object o7, Object o8) {
-
-    }
-
-    @Override
-    public void warn(Marker marker, String s, Object o, Object o1, Object o2, Object o3, Object o4, Object o5, Object o6, Object o7, Object o8, Object o9) {
-
-    }
-
-    @Override
-    public void warn(String s, Object o) {
-
-    }
-
-    @Override
-    public void warn(String s, Object o, Object o1) {
-
-    }
-
-    @Override
-    public void warn(String s, Object o, Object o1, Object o2) {
-
-    }
-
-    @Override
-    public void warn(String s, Object o, Object o1, Object o2, Object o3) {
-
-    }
-
-    @Override
-    public void warn(String s, Object o, Object o1, Object o2, Object o3, Object o4) {
-
-    }
-
-    @Override
-    public void warn(String s, Object o, Object o1, Object o2, Object o3, Object o4, Object o5) {
-
-    }
-
-    @Override
-    public void warn(String s, Object o, Object o1, Object o2, Object o3, Object o4, Object o5, Object o6) {
-
-    }
-
-    @Override
-    public void warn(String s, Object o, Object o1, Object o2, Object o3, Object o4, Object o5, Object o6, Object o7) {
-
-    }
-
-    @Override
-    public void warn(String s, Object o, Object o1, Object o2, Object o3, Object o4, Object o5, Object o6, Object o7, Object o8) {
-
-    }
-
-    @Override
-    public void warn(String s, Object o, Object o1, Object o2, Object o3, Object o4, Object o5, Object o6, Object o7, Object o8, Object o9) {
-
-    }
-}
