--- src/main/resources/ModTheSpire.json
@@ -1,10 +1,0 @@
-{
-  "modid": "${project.artifactId}",
-  "name": "${project.name}",
-  "author_list": ["Board Engineer"],
-  "description": "${project.description}",
-  "version": "${project.version}",
-  "sts_version": "${SlayTheSpire.version}",
-  "mts_version": "${ModTheSpire.version}",
-  "dependencies": ["basemod", "SaveStateMod", "LudicrousSpeed"]
-}
